backend/requirements.txt
Flask     
Flask-Bcrypt     
Flask-Cors     
Flask-Login     
Flask-Session     
Flask-SQLAlchemy
Flask-Migrate
joblib
matplotlib
numpy
pip
python-dotenv
redis
scikit-learn
SQLAlchemy